+++ logprep/util/grok/grok.py
@@ -92,15 +92,15 @@ class Grok:
 
         match_obj = [match for match in match_obj if match is not None]
         matches = [match.groupdict() for match in match_obj]
+        if not matches:
+            return {}
+        first_match = matches[0]
         if self.type_mapper:
-            matches = list(map(self._map_types, matches))
-        # deduplicate matches
-        matches = [dict(tup) for tup in {tuple(match.items()) for match in matches}]
-        return {
-            self.field_mapper[field_hash]: value
-            for match in matches
-            for field_hash, value in match.items()
-        }
+            for key, match in first_match.items():
+                type_ = INT_FLOAT.get(self.type_mapper.get(key))
+                if type_ is not None and match is not None:
+                    first_match[key] = type_(match)
+        return {self.field_mapper[field_hash]: value for field_hash, value in first_match.items()}
 
     def _map_types(self, matches):
         for key, match in matches.items():